education-planner/requirements.txt
fastapi==0.121.2
uvicorn==0.38.0
pydantic==2.12.4
pydantic-settings==2.6.1
python-jose==3.5.0
bcrypt==5.0.0
python-multipart==0.0.20
psycopg2-binary==2.9.11
redis>=5.0.0
httpx==0.28.1
prometheus-client>=0.20.0
slowapi>=0.1.9
email-validator==2.2.0
python-dotenv==1.2.1
passlib==1.7.4
